frontend: extract app lifecycle hooks
